feat|compile: Parallel compile and separate vars
Compilation of multiple datasets can now be done in parallel. By default
it takes either the maximum number of datasets for distributing cpus or
the maximum number of available cpus -1.
Variables such as IceThk, SurfElev and BedElev are now also in separate
files for each trace so that we can get these easy by indexing them with 'age'.