feat: Add region and basin entry from IMBIE
The function get_imbie_basins in post compiling uses the shape files
from Rignot to locate in which region and basin(s) each individual
transect falls in. When it crosses several basins, the indexing_database
function place a pointer to the transect for each basin, so that the
transect is associated to those multiple basins.